+++ packages/react-ui/src/index.d.ts
@@ -22,6 +22,7 @@ import { CHART_TYPES } from './widgets/TimeSeriesWidgetUI/utils/constants';
 import TableWidgetUI from './widgets/TableWidgetUI/TableWidgetUI';
 import NoDataAlert from './widgets/NoDataAlert';
 import FeatureSelectionWidgetUI from './widgets/FeatureSelectionWidgetUI';
+import ComparativeFormulaWidgetUI from './widgets/ComparativeFormulaWidgetUI';
 
 export {
   cartoThemeOptions,
@@ -42,6 +43,7 @@ export {
   TableWidgetUI,
   LegendWidgetUI,
   RangeWidgetUI,
+  ComparativeFormulaWidgetUI,
   LEGEND_TYPES,
   NoDataAlert,
   LegendCategories,